progress

+++ packages/front-end/components/Settings/MemberList.tsx
@@ -1,13 +1,19 @@
 import { FC, useState, useContext } from "react";
 import { FaPlus, FaPencilAlt } from "react-icons/fa";
 import InviteModal from "./InviteModal";
-import { useAuth, MemberRole } from "../../services/auth";
+import { useAuth, MemberRole, MemberStatus } from "../../services/auth";
 import { UserContext } from "../ProtectedPage";
 import DeleteButton from "../DeleteButton";
 import Modal from "../Modal";
 import RoleSelector from "./RoleSelector";
 
-type Member = { id: string; name: string; email: string; role: MemberRole };
+type Member = {
+  id: string;
+  name: string;
+  email: string;
+  role: MemberRole;
+  status: MemberStatus;
+};
 
 const MemberList: FC<{
   members: Member[];
@@ -18,6 +24,7 @@ const MemberList: FC<{
   const { userId } = useContext(UserContext);
   const [roleModal, setRoleModal] = useState<Member>(null);
   const [role, setRole] = useState<MemberRole>("admin");
+  // const [status, setStatus] = useState<MemberStatus>("verified");
 
   const onInvite = () => {
     setInviting(true);
@@ -59,44 +66,51 @@ const MemberList: FC<{
             <th>Email</th>
             <th>Role</th>
             <th />
+            <th />
           </tr>
         </thead>
         <tbody>
-          {members.map((member) => (
-            <tr key={member.id}>
-              <td>{member.name}</td>
-              <td>{member.email}</td>
-              <td>{member.role}</td>
-              <td>
-                {member.id !== userId && (
-                  <>
-                    <a
-                      href="#"
-                      className="tr-hover mr-3"
-                      onClick={(e) => {
-                        e.preventDefault();
-                        setRoleModal(member);
-                        setRole(member.role);
-                      }}
-                    >
-                      <FaPencilAlt />
-                    </a>
-                    <DeleteButton
-                      link={true}
-                      className="tr-hover"
-                      displayName={member.email}
-                      onClick={async () => {
-                        await apiCall(`/member/${member.id}`, {
-                          method: "DELETE",
-                        });
-                        mutate();
-                      }}
-                    />
-                  </>
-                )}
-              </td>
-            </tr>
-          ))}
+          {members.map((member) => {
+            if (member.status === "unverified") {
+              return;
+            }
+            return (
+              <tr key={member.id}>
+                <td>{member.name}</td>
+                <td>{member.email}</td>
+                <td>{member.role}</td>
+                <td>{member.status === "unverified" && "unverified"}</td>
+                <td>
+                  {member.id !== userId && (
+                    <>
+                      <a
+                        href="#"
+                        className="tr-hover mr-3"
+                        onClick={(e) => {
+                          e.preventDefault();
+                          setRoleModal(member);
+                          setRole(member.role);
+                        }}
+                      >
+                        <FaPencilAlt />
+                      </a>
+                      <DeleteButton
+                        link={true}
+                        className="tr-hover"
+                        displayName={member.email}
+                        onClick={async () => {
+                          await apiCall(`/member/${member.id}`, {
+                            method: "DELETE",
+                          });
+                          mutate();
+                        }}
+                      />
+                    </>
+                  )}
+                </td>
+              </tr>
+            );
+          })}
         </tbody>
       </table>
       <button className="btn btn-success mt-3" onClick={onInvite}>
